requirements.txt
# ── Core ──
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0

# ── Database ──
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.0

# ── Cache / Queue ──
redis>=5.0.0
aiohttp>=3.9.0

# ── Scheduling ──
apscheduler>=3.10.0

# ── AI / LLM ──
openai>=1.12.0
google-genai>=0.5.0
groq>=0.4.2

# ── Scraping ──
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.1.0

# ── Data ──
pandas>=2.2.0
yfinance>=0.2.36

# ── Delivery ──
python-telegram-bot>=21.0
aiosmtplib>=3.0.0
jinja2>=3.1.0

# ── Utilities ──
python-dotenv>=1.0.0
loguru>=0.7.0

# ── Testing ──
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
